requirements.txt
tensorflow == 2.1.0
pandas == 1.0.3
xlrd == 1.2.0
jieba == 0.42.1
numpy == 1.18.2
xlsxwriter == 1.2.8
graia-application-mirai == 0.16.1
graia-broadcast == 0.7.0